+++ src/concha/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+from .product import *
+from .model import *
+from .planner import Planner
 from pkg_resources import get_distribution, DistributionNotFound
 
 try: